Add proof of concept CMake build system
CMake tends to be the standard build system for C++ development. Many
projects support CMake and thus importing them into another project
becomes as simple as simply including the ProjectConfig.cmake file.

As mentioned in issue #24, a python interface via pybind11 would make
this toolset orders of magnitude easier to use and manipulatable via
the plentiful tools for data science available for python. Including
pybind11 is as simple as copy and pasting it into this directory and
adding `include_directory(pybind11)`.